+++ src/main/java/org/springframework/aop/PointcutAdvisor.java
@@ -1,0 +1,10 @@
+package org.springframework.aop;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/11/9
+ * @project: sbs-small-spring
+ */
+public interface PointcutAdvisor extends Advisor {
+  Pointcut getPointcut();
+}